% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildwright-package.R
\docType{package}
\name{buildwright-package}
\alias{buildwright}
\alias{buildwright-package}
\title{buildwright: Dependency-Health and Build-Failure Diagnostics}
\description{
buildwright scans installed libraries and \code{renv} lockfiles, builds a
dependency graph, detects version conflicts and cycles, maps packages to
their system requirements, and simulates installation order so that build
failures can be caught \emph{before} an install is attempted. See
\code{\link[=scan_library]{scan_library()}}, \code{\link[=build_dep_graph]{build_dep_graph()}}, \code{\link[=detect_conflicts]{detect_conflicts()}},
\code{\link[=check_sysreqs]{check_sysreqs()}}, \code{\link[=simulate_install]{simulate_install()}}, \code{\link[=diagnose]{diagnose()}} and \code{\link[=report]{report()}}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ashenoy/buildwright}
  \item \url{https://ashenoy.github.io/buildwright/}
  \item Report bugs at \url{https://github.com/ashenoy/buildwright/issues}
}

}
\author{
\strong{Maintainer}: Amit Shenoy \email{shenoy.am@husky.neu.edu}

Authors:
\itemize{
  \item Amit Shenoy \email{shenoy.am@husky.neu.edu}
}

}
\keyword{internal}
